Move E2EE components from friends/ into their own identity module
Six of twelve files under components/friends/ were identity, not friends: none
imported friends.store, all six import identity.store, and their consumers are
Account settings and onboarding rather than any friends view. KeyRestoreDialog
moves too — it fires on any device needing a seed import, not just first run.

The friends barrel had one consumer and dragged the whole identity dialog tree
into Lookout, so it is replaced with direct imports.
