--- web/src/components/friends/index.ts
@@ -1,7 +1,0 @@
-export { default as FriendCard } from './FriendCard.vue'
-export { default as FriendsList } from './FriendsList.vue'
-export { default as FriendInvitations } from './FriendInvitations.vue'
-export { default as AddFriendDialog } from './AddFriendDialog.vue'
-export { default as RecoveryKeyDialog } from './RecoveryKeyDialog.vue'
-export { default as IdentitySettings } from './IdentitySettings.vue'
-
