+++ web/src/components/identity/PasskeyOfferContent.vue
