+++ web/src/components/identity/RecoveryKeySetupContent.vue
