+++ web/src/components/identity/RotateKeysDialog.vue
