+++ web/src/components/onboarding/RecoveryKeyStep.vue
@@ -3,7 +3,7 @@ import { ref, inject, onMounted, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useIdentityStore } from '@/stores/identity.store'
 import { validateKey } from './types'
-import RecoveryKeySetupContent from '@/components/friends/RecoveryKeySetupContent.vue'
+import RecoveryKeySetupContent from '@/components/identity/RecoveryKeySetupContent.vue'
 
 const { t } = useI18n()
 const identityStore = useIdentityStore()
